+++ start_worklog.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 # Worklog Manager - Unix/Linux/Mac Startup Script
-# Version 2.0.0
+# Version 1.5.0
 #
 # This shell script provides an easy way to start Worklog Manager on Unix-like systems